# [Non-breaking changes] Update RLP library and RLP encoder code generation tool for structs