--- a/rlp/rlpgen/testdata/rawvalue.in.txt
+++ b/rlp/rlpgen/testdata/rawvalue.in.txt
@@ -1,0 +1,11 @@
+// -*- mode: go -*-
+
+package test
+
+import "github.com/tomochain/tomochain/rlp"
+
+type Test struct {
+	RawValue          rlp.RawValue
+	PointerToRawValue *rlp.RawValue
+	SliceOfRawValue   []rlp.RawValue
+}